Move YAML encoding support code to pkg/encoding.
