--- cmd/registry/patch/yaml.go
@@ -1,39 +1,0 @@
-// Copyright 2022 Google LLC. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package patch
-
-import (
-	"bytes"
-	"io"
-
-	"gopkg.in/yaml.v3"
-)
-
-// Prefer this encoder because it uses tighter 2-space indentation.
-func yamlEncoder(dst io.Writer) *yaml.Encoder {
-	enc := yaml.NewEncoder(dst)
-	enc.SetIndent(2)
-	return enc
-}
-
-// Encode a patch model.
-func Encode(model interface{}) ([]byte, error) {
-	var b bytes.Buffer
-	err := yamlEncoder(&b).Encode(model)
-	if err != nil {
-		return nil, err
-	}
-	return b.Bytes(), nil
-}
